backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LangGraph + OpenAI
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.40.0

# Vector store (use 'pinecone', not 'pinecone-client')
pinecone>=5.0.0

# Web search
tavily-python>=0.5.0

# Config & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
aiofiles>=24.1.0
httpx>=0.27.0

# Diff parsing (optional; can use stdlib)
unidiff>=0.7.0
